requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
pillow==10.4.0
PyJWT==2.9.0
sqlparse==0.5.1
